+++ server/controllers/donationCtrl.js
@@ -6,7 +6,6 @@ getDonations:async(req, res)=>{
         // let id = req.session.user ? userId : 0;
         let id = +req.params.id;
         const donations = await db.donations.getDonations(id);
-        console.log(req.params.id)
         res.status(200).json(donations)
 },
 
@@ -37,7 +36,6 @@ postFavourite: function (req,res){
 deleteDonation:  function (req, res) {
     const db = req.app.get('db')
         const donation_id = +req.params.id
-        console.log(donation_id)
         db.donations.deleteDonation(donation_id)
     .then(response =>{
         console.log(donation_id)
@@ -66,31 +64,26 @@ updateViewCount: function (req, res) {
     getDonationByCategory:async (req,res)=>{
         const db = req.app.get('db');
         const {category} = req.query;
-        // console.log(category)
         const donationsByCategory = await db.donations.getDonationsByCategory(category);
         res.status(200).json(donationsByCategory);
-        // console.log(donationsByCategory);
     },
     getFilteredDonations:async(req,res)=>{
         const db =req.app.get('db');
         const {title} = req.query;
         const filteredDonations = await db.donations.getDonationsBySearch(title);
         res.status(200).json(filteredDonations);
-        // console.log(filteredDonations)
     },
     getDonationPhotos: async (req,res)=>{
         const db = req.app.get('db');
         const donation_id = +req.params.id;
         const donationPhotos = await db.donations.getDonationPhotos(donation_id);
         res.status(200).json(donationPhotos);
-        // console.log(donationPhotos);
     },
     getDonationInfo: async (req,res)=>{
         const db = req.app.get('db');
         const donation_id = +req.params.id;
         const donationInfo = await db.donations.getDonationInfo(donation_id);
         res.status(200).json(donationInfo);
-        // console.log(donationInfo);
     },
     getUserFavorites: async (req, res) => {
         const db = req.app.get('db');
