+++ src/components/Profile/Profile.js
@@ -1,169 +1,41 @@
 import React from "react";
 import { connect } from "react-redux";
-import { editUser } from "../../redux/reducers/authReducer";
-import axios from 'axios';
+import { Redirect } from 'react-router-dom';
 import "../stylescomponent/Profile.scss";
+import 'bootstrap/dist/css/bootstrap.min.css';
+import Tabs from 'react-bootstrap/Tabs';
+import Tab from 'react-bootstrap/Tab';
+import Edit from './Tabs/Edit';
+import MyDonations from './Tabs/MyDonations';
+import Favorites from './Tabs/Favorites';
 
 class Profile extends React.Component {
-  constructor() {
-    // let { username, password, loginUser } = props;
-    super();
-    this.state = {
-      username: "",
-      location: "",
-      url: "",
-      password: "",
-      currentPassword: "",
-      newPassword: "",
-      confirmPassword: "",
-    };
-  }
-  //make state here to hold the inputs
-  handleInput = (e) => {
-    this.setState({ [e.target.name]: e.target.value });
-  };
-
-  handleSaveUser = () => {
-      const { username, location, url  } = this.state
-      if (username && location && url) {
-        axios.put('/auth/editUser', {
-            username,
-            location,
-            url,
-        })
-        .then(() => {
-            alert("successfully changed username")
-        })
-        .catch(() => {
-            alert("try again")
-        })
-    } else {
-      alert("not equal combo");
-    }
-  }
-
-  // make method like handle input & handle click
-  handleClick = () => {
-    const { username, location, url } = this.state;
-
-    this.props.editUser({ username, location, url });
-  };
-
-  handleNewPassword = (event) => {
-    this.setState({ newPassword: event.target.value });
-  };
-
-  handleConfirmPassword = (event) => {
-    this.setState({ confirmPassword: event.target.value });
-  };
-
-  handleSavePassword = () => {
-    const { newPassword, confirmPassword, currentPassword } = this.state;
-
-    if (newPassword && confirmPassword && newPassword === confirmPassword) {
-        axios.put('/auth/editPassword', {
-            user: this.props.user,
-            password: currentPassword,
-            newPassword: newPassword
-        })
-        .then(() => {
-            alert("successfully changed password")
-        })
-        .catch((res) => { 
-            alert("server trouble -- try again")
-            console.log(res)
-        })
-    } else {
-      alert("not equal password");
-    }
-  };
 
   render() {
-      console.log(this.props.user)
+    if(!this.props.user.user_id) return <Redirect to='/' />
     return (
       <div className="profile">
-        <div className="profile-header">Tap on photo to change it</div>
-        <div className="profile-content">
-          <div className="userName">
-            <div className="userName-text">New Username</div>
-            <input
-              className="input-username"
-              onChange={this.handleInput}
-              name="username"
-              placeholder="new username"
-            />
-            <div className="userName-text">New Location</div>
-            <input
-              className="input-username"
-              onChange={this.handleInput}
-              name="location"
-              placeholder="new location"
-            />
-            <div className="userName-text">URL</div>
-            <input
-              className="input-username"
-              onChange={this.handleInput}
-              name="url"
-              placeholder="paste image URL"
-            />
-            <div className="button-wrapper">
-              <button className="save-changes" onClick={this.handleSaveUser}>
-                Save changes
-              </button>
-            </div>
-          </div>
-          <div className="passWord">
-            <div className="change-password">Change Password</div>
-            <div className="userName-text">Current Password</div>
-            <input
-              className="input-username"
-              type="password"
-              onChange={this.handleInput}
-              name="currentPassword"
-              placeholder="current password"
-            />
-            <div className="userName-text">New Password</div>
-            <input
-              className="input-username"
-              type="password"
-              onChange={this.handleNewPassword}
-              name="newPassword"
-              placeholder="new password"
-            />
-            <div className="userName-text">Confirm Password</div>
-            <input
-              className="input-username"
-              type="password"
-              onChange={this.handleConfirmPassword}
-              name="confirmPassword"
-              placeholder="confirm password"
-            />
-            <div className="button-wrapper">
-              <button
-                className="save-changes"
-                onClick={this.handleSavePassword}
-              >
-                Save changes
-              </button>
-            </div>
-          </div>
-        </div>
+        <Tabs defaultActiveKey='edit' id='profile-tabs'>
+          <Tab eventKey='edit' title='Edit Profile'>
+            <Edit />
+          </Tab>
+          <Tab eventKey='favorites' title='Favorited'>
+            <Favorites />
+          </Tab>
+          <Tab eventKey='user-donations' title='My Donations'>
+            <MyDonations user={this.props.user} />
+          </Tab>
+        </Tabs>
       </div>
     );
   }
 }
 
-const mapStateToProps = (state) => {
-  let { user, username, password, loading } = state.authReducer;
-  console.log(state);
-  return {
-    user,
-    username,
-    password,
-    loading,
-  };
-};
+const mapStateToProps = state => {
+    let { user } = state.authReducer;
+    return{
+        user
+    }
+}
 
-export default connect(mapStateToProps, {
-  editUser,
-})(Profile);
+export default connect(mapStateToProps)(Profile);
